point to config file, fix css module mapping, config for shim, file rename
