+++ jest.config.json
@@ -1,15 +1,11 @@
 {
   "verbose": true,
 
+  "setupFiles": [
+    "raf/polyfill"
+  ],
+
   "coverageReporters": ["html", "json", "lcov", "cobertura"],
-
-  "testRegex": "(/src/.*\\.(test|spec))\\.(ts|tsx|js|jsx)$",
-
-  "moduleFileExtensions": ["ts", "tsx", "js", "jsx"],
-
-  "transform": {
-    "^.\\.jsx$": "<rootDir>/node_modules/babel-jest"
-  },
 
   "moduleNameMapper": {
     "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/test/__mocks__/empty-module.js",
@@ -19,8 +15,8 @@
   "coverageDirectory": "<rootDir>/reports/test-coverage",
 
   "collectCoverageFrom": [
-    "src/components/**/*.{ts,tsx}",
-    "src/services/**/*.{ts,tsx}",
-    "src/views/**/*.{ts,tsx}"
+    "src/components/**/*.{js,jsx}",
+    "src/services/**/*.{js,jsx}",
+    "src/views/**/*.{js,jsx}"
   ]
 }
