+++ package.json
@@ -6,9 +6,9 @@
   "scripts": {
     "build": "npm run lint && webpack --config ./webpack.config.prod.js -p",
     "develop": "webpack-dev-server --open --config ./webpack.config.develop.js",
-    "lint": "eslint *.js",
+    "lint": "eslint *.js ./src/**/**/*.spec.jsx",
     "serve": "npm run build && http-server ./build -o -p 9000",
-    "test": "jest"
+    "test": "jest --config ./jest.config.json"
   },
   "repository": {
     "url": "https://github.com/ProvidenceGeeks/website-frontend.git",
@@ -50,11 +50,5 @@
     "webpack-dev-server": "^2.9.1",
     "webpack-md5-hash": "^0.0.5",
     "webpack-merge": "^4.1.0"
-  },
-  "jest": {
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/test/__mocks__/empty-module.js",
-      "\\.(css|scss)$": "<rootDir>/test/__mocks__/empty-module.js"
-    }
   }
 }
