+++ src/components/bootstrap/bootstrap.spec.jsx
@@ -1,16 +1,15 @@
 import * as React from 'react';
-import raf from '../../../test/shims/react-raf-shim';
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
-import Home from './home';
+import Bootstrap from './bootstrap';
 
 configure({ adapter: new Adapter() });
 
-describe('Hello View', () => {
-  const home = shallow(<Home />);
+describe('Bootstrap Component', () => {
+  const bootstrap = shallow(<Bootstrap />);
 
   it('should not be null', () => {
-    expect(home).not.toBeNull();
+    expect(bootstrap).not.toBeNull();
   });
 });
